Refactoring and Source API changes
