+++ config.js
@@ -1,0 +1,4 @@
+module.exports = {
+    deathRecoveriesURL : 'https://api.covid19india.org/deaths_recoveries.json',
+    rawDataURL : 'https://api.covid19india.org/raw_data.json'
+}
