+++ package.json
@@ -11,7 +11,7 @@
   "license": "ISC",
   "dependencies": {
     "commander": "^5.0.0",
-    "puppeteer": "^1.3.0",
+    "node-fetch": "^2.6.0",
     "request": "^2.88.2"
   }
 }
